backend/requirements.txt
flask
flask-cors
pymongo
mysql-connector-python
pyodbc
pandas
openpyxl
reportlab
python-dotenv
cryptography
flask-limiter
flask-talisman
python-pptx
openai
